tb3289-teimurazi-bakuradze-winter-2024-001-se-181-course-project-master/src/main/java/banking/PassTimeProcessor.java
package banking;

public class PassTimeProcessor extends CommandProcessor {
    private int timeToPass;

    PassTimeProcessor(Bank bank) {
        super(bank);
    }

    public void processPassTime(String[] commandParts) {
        this.timeToPass = Integer.parseInt(commandParts[1]);
        bank.passTimeForExistingAccounts(timeToPass);
        for (int i = 0; i < timeToPass; i++) {
            bank.closeAccountsWithZeroBalance();
            bank.minimumBalancePayment();
            bank.accrueApr();
        }
    }


}
